Storage/RAW.cpp: Fix harmless parfait warning about a possible integer overflow before assignment, bugref:3409
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@107725 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Storage/RAW.cpp
@@ -467,7 +467,7 @@ static int rawProbeIsIso9660OrUdf(PVDINTERFACEIOINT pIfIo, PVDIOSTORAGE pStorage
                  && MATCH_HDR(&Buf.VolDescHdr, UDF_EXT_VOL_DESC_TYPE, UDF_EXT_VOL_DESC_STD_ID_BOOT, UDF_EXT_VOL_DESC_VERSION) )
         {
             if (offUdfBootVolDesc == UINT64_MAX)
-                offUdfBootVolDesc = iVolDesc * cbSector;
+                offUdfBootVolDesc = (uint64_t)iVolDesc * cbSector;
             else
                 return RTERRINFO_LOG_SET_F(pErrInfo, VERR_VFS_BOGUS_FORMAT, "Only one BOOT2 descriptor is supported");
         }
